Refs #7 - Removed explicit platform specification on build. Changed platform to default. Relying on javac target.
Used to set the active platform explicitly to JDK 1.4, but we're in the year 2012! This should not be a problem anymore. At most the user will have run time link problems on Java 1.4, which is easily recoverable (if we accidentally use latest java features). Also I have started to think that even if Java 1.4 happens to be supported, we can use retroweaver for that. Maybe we should update minimum java version to 1.5.